# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Odoo POS Cafe API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /kitchen/tickets:
    get:
      operationId: listKitchenTickets
      tags: [kitchen]
      summary: List kitchen tickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [TO_COOK, PREPARING, COMPLETED]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Kitchen tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KitchenTicket"
    post:
      operationId: createKitchenTicket
      tags: [kitchen]
      summary: Create kitchen ticket
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKitchenTicketBody"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KitchenTicket"
  /kitchen/tickets/{id}/status:
    put:
      operationId: updateKitchenTicketStatus
      tags: [kitchen]
      summary: Update kitchen ticket status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKitchenTicketStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KitchenTicket"
  /kitchen/tickets/{id}/items/{itemId}:
    put:
      operationId: updateKitchenTicketItem
      tags: [kitchen]
      summary: Mark ticket item as prepared
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateKitchenItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KitchenTicketItem"

  # Reporting
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductCategory:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
      required:
        - id
        - name
        - color
...
    ProductVariant:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        attribute:
          type: string
        value:
          type: string
        extraPrice:
          type: number
        unit:
          type: string
      required:
        - id
        - productId
        - attribute
        - value
        - extraPrice
        - unit
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        categoryId:
          type: string
        category:
          $ref: "#/components/schemas/ProductCategory"
        price:
          type: number
        unit:
          type: string
        tax:
          type: number
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - categoryId
        - price
        - unit
        - tax
        - createdAt
...
    Table:
      type: object
      properties:
        id:
          type: string
        floorId:
          type: string
        tableNumber:
          type: string
        seats:
          type: integer
        active:
          type: boolean
        token:
          type: string
      required:
        - id
        - floorId
        - tableNumber
        - seats
        - active
        - token
...
    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        country:
          type: string
      required:
        - id
        - name
        - country
...
    OrderLine:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        productId:
          type: string
        product:
          $ref: "#/components/schemas/Product"
        qty:
          type: integer
        unitPrice:
          type: number
        tax:
          type: number
        subTotal:
          type: number
        total:
          type: number
        note:
          type: ["string", "null"]
      required:
        - id
        - orderId
        - productId
        - qty
        - unitPrice
        - tax
        - subTotal
        - total
...
    OrderSummary:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        tableId:
          type: ["string", "null"]
        table:
          $ref: "#/components/schemas/Table"
        customerId:
          type: ["string", "null"]
        customer:
          $ref: "#/components/schemas/Customer"
        status:
          type: string
          enum: [DRAFT, PAID]
        createdAt:
          type: string
          format: date-time
        total:
          type: number
      required:
        - id
        - sessionId
        - status
        - createdAt
        - total
...
    KitchenTicketItem:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        orderLineId:
          type: string
        orderLine:
          $ref: "#/components/schemas/OrderLine"
        prepared:
          type: boolean
      required:
        - id
        - ticketId
        - orderLineId
        - prepared
    KitchenTicket:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        order:
          $ref: "#/components/schemas/OrderSummary"
        status:
          type: string
          enum: [TO_COOK, PREPARING, COMPLETED]
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/KitchenTicketItem"
      required:
        - id
        - orderId
        - status
        - createdAt
    CreateKitchenTicketBody:
      type: object
      properties:
        orderId:
          type: string
      required:
        - orderId
    UpdateKitchenTicketStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [TO_COOK, PREPARING, COMPLETED]
      required:
        - status
    UpdateKitchenItemBody:
      type: object
      properties:
        prepared:
          type: boolean
      required:
        - prepared
...